Convert tests from TablePropertiesProviderIT to unit tests
